Split out views/system.py
